FB Login
I’m working on adding FB login functionality.  Currently the unity
simulator works fine, but I can’t login from my iOS device still.

There are also several art updates.  Background randomizer. “building
scanning” background on maps level, and the early makings of the
building populator- which creates objects from prefab, but I can’t seem
to translate their position correctly.  Getting the names read from
json onto prefabs was where I gave up.

MY TODO:
-turn “bite” event into a decision point “kill your friend now” vs
“roll to save them” - in which case they need to become injured and
stay in the “inactive” for several days (possibly also consuming
caretakers for the injured).
-Create new Player class that gives some stat variation, possibly tied
into generic FB data (gender age-?)
-Create Weapon class that gives some variation to the modifiers
provided by different classes of weapons.  Also need Ammo.cs Class to
handle what bullets the player has for which gun.
-Create new 4-5 player RPG combat system. Match Weapon.cs onto
Player.cs to calculate damage.  New animator system to handle
auto-attack mode. speed variable to manage swing timers. prep for new
animation art (one day…)
-Remove start screen entirely.  redo menu layout on map screen.
-FB should provide persistent login, so re-opening the app goes
directly to map screen.
-Finding people/loot should dynamically react to the relative #days
alive.  Hour 1-3 will have far more to find than day 100. Manually map
the curve? how to?
-places api maps nearby buildings to map screen
-Story.cs that downloads images from the server, and places them in a
layers that players swipe through (like swappable comic book cells)
(This will be how we will start creating the tutorial section in
future, and more depth to player story beyond that).
-and many more to come…

+++ Assets/Scripts/StartMenuManager.cs
@@ -19,7 +19,7 @@ public class StartMenuManager : MonoBehaviour {
 	}
 
 	void UpdateTheText () {
-		daysSurvived.text = ( Mathf.Floor (GameManager.instance.daysSurvived)).ToString();
+		daysSurvived.text = ((int)(Mathf.Floor (GameManager.instance.daysSurvived))).ToString();
 		activeSurvivors.text = "Active Survivors: " + GameManager.instance.survivorsActive.ToString();
 		membersAlive.text = "Total Living Survivors: " + GameManager.instance.totalSurvivors.ToString();
 		supply.text = "Supply: " + GameManager.instance.supply.ToString();
@@ -37,6 +37,8 @@ public class StartMenuManager : MonoBehaviour {
         int food = GameManager.instance.foodCount;
         int water = GameManager.instance.waterCount;
         
+        //Debug.Log ("meals of food left, should be: "Mathf.Floor( food / GameManager.instance.totalSurvivors));
+        //Debug.Log ("meals of water left, should be: "Mathf.Floor( water / GameManager.instance.totalSurvivors));
         int foodExpireInMeals = (int)Mathf.Floor( food / GameManager.instance.totalSurvivors);
         int waterExpireInMeals = (int)Mathf.Floor( water / GameManager.instance.totalSurvivors);
